Comment out `Cargo.lock` from `Rust.gitignore`
Suggestion to explicitly not check in `Cargo.lock` was removed July 19, 2023, see
https://github.com/rust-lang/cargo/commit/57aa352da74b49edd6243a9cb57df2cd5f51b52b.
It merely states that relying upon it can give a false sense of security and may be
a source of merge conflicts. However current suggestion now seems to be to automate
dependency checking with tools like Dependabot, see
https://blog.rust-lang.org/2023/08/29/committing-lockfiles.html.

This change favors build consistency.